Factor out metadata test, remove dead code
